# packages/inference-api/src/app.ts
import { apiKeyAuthorization,jsonObjectSchema,openApiDocument,requireScope,type ApiKeyResolver,type Job,type JobRepository,type RouteSpec } from '@ai-platform/common';
import { Hono } from 'hono';
import { streamSSE } from 'hono/streaming';
import { readFileSync } from 'node:fs';

const submitSchema = { type: 'object', required: ['idempotencyKey'], properties: { idempotencyKey: { type: 'string', minLength: 1 }, input: jsonObjectSchema } };
function jobStateCounts(jobs: Array<{state:string}>) { return jobs.reduce<Record<string,number>>((counts,job)=>{counts[job.state]=(counts[job.state]??0)+1;return counts;},{}); }
function runtimeMetrics(){let mode='unknown',active=0;try{mode=JSON.parse(readFileSync(process.env.AI_FACTORY_MODE_FILE!,'utf8')).mode;}catch{}try{active=Number(JSON.parse(readFileSync(process.env.AI_RUNTIME_STATUS!,'utf8')).active??0);}catch{}return[`ai_factory_mode{product="inference",mode="${mode}"} 1`,`ai_inference_active_requests ${active}`];}
function internalImportToken(){try{return readFileSync(process.env.ARTIFACT_IMPORT_TOKEN_FILE!,'utf8').trim();}catch{return process.env.ARTIFACT_IMPORT_TOKEN??'';}}
export const inferenceRoutes: RouteSpec[] = [
	{ method:'GET',path:'/healthz',summary:'Liveness' },{ method:'GET',path:'/readyz',summary:'Readiness' },{ method:'GET',path:'/v1/version',summary:'Version',scope:'inference:read' },
	{ method:'GET',path:'/v1/jobs',summary:'List jobs',scope:'jobs:read' },{ method:'GET',path:'/v1/jobs/:id',summary:'Get job',scope:'jobs:read' },{ method:'GET',path:'/v1/jobs/:id/events',summary:'Get job events',scope:'jobs:read' },{ method:'POST',path:'/v1/jobs/:id/cancel',summary:'Cancel job',scope:'jobs:write' },
	{ method:'GET',path:'/v1/jobs/:id/events/stream',summary:'Stream job events',scope:'jobs:read' },{ method:'GET',path:'/v1/events/stream',summary:'Stream job changes',scope:'jobs:read' },{ method:'GET',path:'/v1/metrics',summary:'Prometheus metrics',scope:'metrics:read' },
	{ method:'GET',path:'/v1/models',summary:'List model sources',scope:'inference:read' },{ method:'POST',path:'/v1/adapters/import',summary:'Import signed adapter',scope:'adapters:write',requestSchema:submitSchema },{ method:'GET',path:'/v1/adapters/:id',summary:'Get adapter',scope:'inference:read' },
	{ method:'POST',path:'/v1/evaluations',summary:'Evaluate candidates',scope:'evaluations:write',requestSchema:submitSchema },{ method:'POST',path:'/v1/rankings',summary:'Rank evaluation results',scope:'evaluations:write',requestSchema:submitSchema },
	{ method:'POST',path:'/v1/promotions',summary:'Promote candidate',scope:'deployments:write',requestSchema:submitSchema },{ method:'GET',path:'/v1/deployments/current',summary:'Current deployment',scope:'inference:read' },{ method:'POST',path:'/v1/deployments/rollback',summary:'Rollback deployment',scope:'deployments:write',requestSchema:submitSchema },
	{ method:'GET',path:'/v1/library-deployments',summary:'List library deployment aliases',scope:'inference:read' },{ method:'GET',path:'/v1/library-deployments/:libraryId',summary:'Get library deployment alias',scope:'inference:read' },{ method:'POST',path:'/v1/library-deployments/:libraryId/rollback',summary:'Rollback one library alias',scope:'deployments:write',requestSchema:submitSchema },
];

export const inferenceOpenApi = (version = '0.11.0') => openApiDocument({ title:'AI Inference Control API',version,routes:inferenceRoutes,operationNamespace:'inference' });

export function createInferenceControlApp(input: { jobs: JobRepository; resolveKey: ApiKeyResolver; version?: string; ready?: () => Promise<boolean>; importArtifact?:(job:Job)=>Promise<string>; currentDeployment?:()=>Promise<unknown>; libraryDeployments?:(libraryId?:string)=>Promise<unknown[]>; candidate?:(id:string)=>Promise<unknown|null> }) {
	const app = new Hono(); const version=input.version??'0.11.0'; const document=inferenceOpenApi(version);
	app.get('/healthz',(context)=>context.json({ok:true,service:'inference-control'})); app.get('/readyz',async(context)=>{const ok=await(input.ready?.()??true);return context.json({ok},ok?200:503);});
	app.get('/openapi.json',(context)=>context.json(document)); app.get('/docs',(context)=>context.html('<!doctype html><title>AI Inference API</title><script id="api-reference" data-url="/openapi.json"></script><script src="https://cdn.jsdelivr.net/npm/@scalar/api-reference"></script>'));
	app.post('/internal/artifacts/import',async context=>{if(!input.importArtifact||context.req.header('authorization')!==`Bearer ${internalImportToken()}`)return context.json({error:{code:'not_found',message:'Route not found.'}},404);const job=await context.req.json()as Job;return context.json({resultManifest:await input.importArtifact(job)});});
	app.use('/v1/*',apiKeyAuthorization(input.resolveKey));
	app.get('/v1/version',requireScope('inference:read'),(context)=>context.json({version,api:'v1'}));
	app.get('/v1/jobs',requireScope('jobs:read'),async(context)=>context.json({items:await input.jobs.list(Number(context.req.query('limit')??100))}));
	app.get('/v1/jobs/:id',requireScope('jobs:read'),async(context)=>{const job=await input.jobs.get(context.req.param('id'));return job?context.json(job):context.json({error:{code:'not_found',message:'Job not found.'}},404);});
	app.get('/v1/jobs/:id/events',requireScope('jobs:read'),async(context)=>context.json({items:await input.jobs.events(context.req.param('id'))}));
	app.get('/v1/jobs/:id/events/stream',requireScope('jobs:read'),(context)=>streamSSE(context,async stream=>{let cursor=Number(context.req.header('last-event-id')??0);while(!stream.closed){const events=await input.jobs.events(context.req.param('id'));for(const event of events.filter(({id})=>id>cursor)){await stream.writeSSE({id:String(event.id),event:event.type,data:JSON.stringify(event)});cursor=event.id;}await stream.writeSSE({event:'heartbeat',data:'{}'});await stream.sleep(2000);}}));
	app.get('/v1/events/stream',requireScope('jobs:read'),(context)=>streamSSE(context,async stream=>{let cursor=context.req.header('last-event-id')??'',prior='';while(!stream.closed){const jobs=await input.jobs.list(100),value=JSON.stringify(jobs.map(({id,state,progress,updatedAt})=>({id,state,progress,updatedAt}))),id=String(Math.max(0,...jobs.map(({updatedAt})=>Date.parse(updatedAt))));if(value!==prior&&id!==cursor){await stream.writeSSE({id,event:'jobs',data:value});cursor=id;prior=value;}else await stream.writeSSE({event:'heartbeat',data:'{}'});await stream.sleep(2000);}}));
	app.get('/v1/metrics',requireScope('metrics:read'),async context=>{const jobs=await input.jobs.list(1000),counts=Object.entries(jobStateCounts(jobs)).map(([state,count])=>`ai_jobs{product="inference",state="${state}"} ${count}`);return context.text(`${[...counts,...runtimeMetrics()].join('\n')}\n`,200,{'content-type':'text/plain; version=0.0.4'});});
	app.post('/v1/jobs/:id/cancel',requireScope('jobs:write'),async(context)=>{const job=await input.jobs.cancel(context.req.param('id'));return job?context.json(job):context.json({error:{code:'not_found',message:'Job not found.'}},404);});
	app.get('/v1/models',requireScope('inference:read'),async(context)=>context.json({items:[{id:process.env.SOURCE_MODEL??'Qwen/Qwen3.5-4B',kind:'base'},...(await input.libraryDeployments?.()??[]).map((item:any)=>({id:item.modelAlias,kind:'library-adapter',libraryId:item.libraryId,candidateId:item.candidateId}))]})); app.get('/v1/adapters/:id',requireScope('inference:read'),async(context)=>{const candidate=await input.candidate?.(context.req.param('id'));return candidate?context.json(candidate):context.json({error:{code:'not_found',message:'Adapter candidate not found.'}},404);}); app.get('/v1/deployments/current',requireScope('inference:read'),async(context)=>context.json({deployment:await input.currentDeployment?.()??null}));
	app.get('/v1/library-deployments',requireScope('inference:read'),async context=>context.json({items:await input.libraryDeployments?.()??[]}));app.get('/v1/library-deployments/:libraryId',requireScope('inference:read'),async context=>{const value=(await input.libraryDeployments?.(context.req.param('libraryId'))??[])[0];return value?context.json(value):context.json({error:{code:'not_found',message:'Library deployment not found.'}},404);});
	const submit=(type:string,scope:string,validate?:(request:any)=>string|undefined,extra:Record<string,string>={},maxAttempts?:number)=>[requireScope(scope),async(context:any)=>{const body=await context.req.json();if(!body.idempotencyKey)return context.json({error:{code:'invalid_request',message:'idempotencyKey is required.'}},400);const request={...(body.input??{}),...Object.fromEntries(Object.entries(extra).map(([key,param])=>[key,context.req.param(param)]))};const diagnostic=validate?.(request);if(diagnostic)return context.json({error:{code:'invalid_request',message:diagnostic}},400);const job=await input.jobs.submit({type,request,idempotencyKey:body.idempotencyKey,priority:body.priority,maxAttempts});return context.json(job,202);} ] as const;
	app.post('/v1/adapters/import',...submit('adapter.import','adapters:write',(request)=>request.sourceId!=='training-local'||typeof request.manifestUri!=='string'?'sourceId training-local and manifestUri are required.':undefined,{},8)); app.post('/v1/evaluations',...submit('evaluation.run','evaluations:write')); app.post('/v1/rankings',...submit('ranking.run','evaluations:write')); app.post('/v1/promotions',...submit('deployment.promote','deployments:write',(request)=>typeof request.candidateId!=='string'||typeof request.rankingManifest!=='string'?'candidateId and rankingManifest are required.':undefined)); app.post('/v1/deployments/rollback',...submit('deployment.rollback','deployments:write'));
	app.post('/v1/library-deployments/:libraryId/rollback',...submit('deployment.rollback','deployments:write',undefined,{libraryId:'libraryId'}));
	app.notFound((context)=>context.json({error:{code:'not_found',message:'Route not found.'}},404)); app.onError((error,context)=>context.json({error:{code:'internal_error',message:error.message}},500)); return app;
}