# packages/treeai-sdk/openapi/inference.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "AI Inference Control API",
    "version": "0.11.0"
  },
  "paths": {
...
    "/v1/jobs": {
      "get": {
        "operationId": "inference.get.jobs",
        "summary": "List jobs",
        "security": [
          {
            "apiKey": [
              "jobs:read"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "202": {
            "description": "Accepted"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/v1/jobs/{id}": {
      "get": {
        "operationId": "inference.get.jobs.by-id",
        "summary": "Get job",
        "security": [
          {
            "apiKey": [
              "jobs:read"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "202": {
            "description": "Accepted"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/v1/jobs/{id}/events": {
      "get": {
        "operationId": "inference.get.jobs.by-id.events",
        "summary": "Get job events",
        "security": [
          {
            "apiKey": [
              "jobs:read"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "202": {
            "description": "Accepted"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/v1/jobs/{id}/cancel": {
      "post": {
        "operationId": "inference.post.jobs.by-id.cancel",
        "summary": "Cancel job",
        "security": [
          {
            "apiKey": [
              "jobs:write"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "202": {
            "description": "Accepted"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/v1/jobs/{id}/events/stream": {
      "get": {
        "operationId": "inference.get.jobs.by-id.events.stream",
        "summary": "Stream job events",
        "security": [
          {
            "apiKey": [
              "jobs:read"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "202": {
            "description": "Accepted"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/v1/events/stream": {
      "get": {
        "operationId": "inference.get.events.stream",
        "summary": "Stream job changes",
        "security": [
          {
            "apiKey": [
              "jobs:read"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "202": {
            "description": "Accepted"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
    "/v1/models": {
      "get": {
        "operationId": "inference.get.models",
        "summary": "List model sources",
        "security": [
          {
            "apiKey": [
              "inference:read"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "202": {
            "description": "Accepted"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/v1/adapters/import": {
      "post": {
        "operationId": "inference.post.adapters.import",
        "summary": "Import signed adapter",
        "security": [
          {
            "apiKey": [
              "adapters:write"
            ]
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "idempotencyKey"
                ],
                "properties": {
                  "idempotencyKey": {
                    "type": "string",
                    "minLength": 1
                  },
                  "input": {
                    "type": "object",
                    "additionalProperties": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "202": {
            "description": "Accepted"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/v1/adapters/{id}": {
      "get": {
        "operationId": "inference.get.adapters.by-id",
        "summary": "Get adapter",
        "security": [
          {
            "apiKey": [
              "inference:read"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "202": {
            "description": "Accepted"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/v1/evaluations": {
      "post": {
        "operationId": "inference.post.evaluations",
        "summary": "Evaluate candidates",
        "security": [
          {
            "apiKey": [
              "evaluations:write"
            ]
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "idempotencyKey"
                ],
                "properties": {
                  "idempotencyKey": {
                    "type": "string",
                    "minLength": 1
                  },
                  "input": {
                    "type": "object",
                    "additionalProperties": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "202": {
            "description": "Accepted"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/v1/rankings": {
      "post": {
        "operationId": "inference.post.rankings",
        "summary": "Rank evaluation results",
        "security": [
          {
            "apiKey": [
              "evaluations:write"
            ]
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "idempotencyKey"
                ],
                "properties": {
                  "idempotencyKey": {
                    "type": "string",
                    "minLength": 1
                  },
                  "input": {
                    "type": "object",
                    "additionalProperties": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "202": {
            "description": "Accepted"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/v1/promotions": {
      "post": {
        "operationId": "inference.post.promotions",
        "summary": "Promote candidate",
        "security": [
          {
            "apiKey": [
              "deployments:write"
            ]
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "idempotencyKey"
                ],
                "properties": {
                  "idempotencyKey": {
                    "type": "string",
                    "minLength": 1
                  },
                  "input": {
                    "type": "object",
                    "additionalProperties": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "202": {
            "description": "Accepted"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/v1/deployments/current": {
      "get": {
        "operationId": "inference.get.deployments.current",
        "summary": "Current deployment",
        "security": [
          {
            "apiKey": [
              "inference:read"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "202": {
            "description": "Accepted"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/v1/deployments/rollback": {
      "post": {
        "operationId": "inference.post.deployments.rollback",
        "summary": "Rollback deployment",
        "security": [
          {
            "apiKey": [
              "deployments:write"
            ]
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "idempotencyKey"
                ],
                "properties": {
                  "idempotencyKey": {
                    "type": "string",
                    "minLength": 1
                  },
                  "input": {
                    "type": "object",
                    "additionalProperties": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "202": {
            "description": "Accepted"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/v1/library-deployments": {
      "get": {
        "operationId": "inference.get.library-deployments",
        "summary": "List library deployment aliases",
        "security": [
          {
            "apiKey": [
              "inference:read"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "202": {
            "description": "Accepted"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/v1/library-deployments/{libraryId}": {
      "get": {
        "operationId": "inference.get.library-deployments.by-libraryId",
        "summary": "Get library deployment alias",
        "security": [
          {
            "apiKey": [
              "inference:read"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "202": {
            "description": "Accepted"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/v1/library-deployments/{libraryId}/rollback": {
      "post": {
        "operationId": "inference.post.library-deployments.by-libraryId.rollback",
        "summary": "Rollback one library alias",
        "security": [
          {
            "apiKey": [
              "deployments:write"
            ]
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "idempotencyKey"
                ],
                "properties": {
                  "idempotencyKey": {
                    "type": "string",
                    "minLength": 1
                  },
                  "input": {
                    "type": "object",
                    "additionalProperties": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "202": {
            "description": "Accepted"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "ak_<id>_<secret>"
      }
    }
  }
}